# back-end/API/app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Gestion de demandes d'emplois",
    "version": "1.0.0"
  },
  "paths": {
    "/ping": {
      "get": {
        "description": "Checks if the server is up and running",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Connected successfully"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "Logs in a user",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "User information",
            "in": "body",
            "description": "JSON data containing user information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User created successfully"
          },
          "401": {
            "description": "User doesn't exist"
          }
        }
      }
    },
    "/createUser": {
      "post": {
        "description": "Creates a user",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "User information",
            "in": "body",
            "description": "JSON data containing user information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Invalid data"
          }
        }
      }
    },
    "/updatePassword": {
      "put": {
        "description": "Updates a user's password",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "User information",
            "in": "body",
            "description": "JSON data containing user information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password updated successfully"
          },
          "400": {
            "description": "Invalid data"
          }
        }
      }
    },
    "/getAllUsers": {
      "get": {
        "description": "Get a list of all users",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Users successfully fetched"
          }
        }
      }
    },
    "/getUser": {
      "get": {
        "description": "Get a user's information based on email address",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "User information",
            "in": "body",
            "description": "JSON data containing user information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully fetched"
          },
          "400": {
            "description": "Missing email address"
          }
        }
      }
    }
  }
}